Big commit: - camera system - parallel task - sphere test

+++ Renderer/include/functional/utils/image/image.h
@@ -10,26 +10,27 @@ namespace functional
 
 /**
  * @brief Abstraction of Image. Need to use ImageGenerator to persist.
- * 
+ *
  */
 class Image
 {
   public:
-    Image(size_t width, size_t height);
+    Image(uint32_t width, uint32_t height);
     Image(const Image& image);
     Image(Image&& image) = delete;
     Image& operator=(const Image& image);
     Image& operator=(Image&& image) = delete;
     ~Image();
 
-    void set_pixel(size_t x, size_t y, float r, float g, float b);
-    void set_pixel(size_t x, size_t y, const Color& color);
-    Color get_pixel(size_t x, size_t y) const;
-    size_t get_width() const;
-    size_t get_height() const;
+    void set_pixel(uint32_t x, uint32_t y, float r, float g, float b);
+    void set_pixel(uint32_t x, uint32_t y, const Color& color);
+    Color get_pixel(uint32_t x, uint32_t y) const;
+    uint32_t get_width() const;
+    uint32_t get_height() const;
+
   private:
-    size_t width, height;
-    std::vector<Color>   pixels;
+    uint32_t width, height;
+    std::vector<Color> pixels;
 };
 } // namespace functional
 } // namespace huan_renderer_cpu
